+++ core/src/main/java/nl/nn/adapterframework/jdbc/dbms/DbmsSupportFactory.java
@@ -1,5 +1,5 @@
 /*
-   Copyright 2013, 2018 Nationale-Nederlanden
+   Copyright 2013, 2018, 2019 Nationale-Nederlanden
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -35,59 +35,69 @@ public class DbmsSupportFactory implements IDbmsSupportFactory {
 
 	private final static String PRODUCT_NAME_ORACLE_="Oracle";
 	private final static String PRODUCT_NAME_MSSQLSERVER="Microsoft SQL Server";
+	private final static String PRODUCT_NAME_MYSQL="MySQL";
+	private final static String PRODUCT_NAME_MARIADB="MariaDB";
 
 	private Properties dbmsSupportMap; 
 
 	public IDbmsSupport getDbmsSupport(Connection conn) {
-		String product;
+		String productName;
+		String productVersion;
 		try {
 			DatabaseMetaData md = conn.getMetaData();
-			product=md.getDatabaseProductName();
+			productName = md.getDatabaseProductName();
+			productVersion = md.getDatabaseProductVersion();
+			if (PRODUCT_NAME_MYSQL.equals(productName) && StringUtils.contains(productVersion, PRODUCT_NAME_MARIADB)) {
+				productName = PRODUCT_NAME_MARIADB;
+			}
 		} catch (SQLException e1) {
 			throw new RuntimeException("cannot obtain product from connection metadata", e1);
 		}
 		Properties supportMap=getDbmsSupportMap();
 		if (supportMap!=null) {
-			if (StringUtils.isEmpty(product)) {
+			if (StringUtils.isEmpty(productName)) {
 				log.warn("no product found from connection metadata");
 			} else {
-				if (!supportMap.containsKey(product)) {
-					log.warn("product ["+product+"] not configured in dbmsSupportMap");
+				if (!supportMap.containsKey(productName)) {
+					log.warn("product ["+productName+"] not configured in dbmsSupportMap");
 				} else {
-					String dbmsSupportClass=supportMap.getProperty(product);
+					String dbmsSupportClass=supportMap.getProperty(productName);
 					if (StringUtils.isEmpty(dbmsSupportClass)) {
-						log.warn("product ["+product+"] configured empty in dbmsSupportMap");
+						log.warn("product ["+productName+"] configured empty in dbmsSupportMap");
 					} else {
 						try {
-							if (log.isDebugEnabled()) log.debug("creating dbmsSupportClass ["+dbmsSupportClass+"] for product ["+product+"]");
+							if (log.isDebugEnabled()) log.debug("creating dbmsSupportClass ["+dbmsSupportClass+"] for product ["+productName+"]");
 							return (IDbmsSupport)ClassUtils.newInstance(dbmsSupportClass);
 						} catch (Exception e) {
-							throw new RuntimeException("Cannot create dbmsSupportClass ["+dbmsSupportClass+"] for product ["+product+"]",e);
+							throw new RuntimeException("Cannot create dbmsSupportClass ["+dbmsSupportClass+"] for product ["+productName+"]",e);
 						} 
 					}
 				}
-			}
-		}
-		else {
+			}				
+		} else {
 			log.warn("no dbmsSupportMap specified, reverting to built in types");
-			if (PRODUCT_NAME_ORACLE_.equals(product)) {
+			if (PRODUCT_NAME_ORACLE_.equals(productName)) {
 				log.debug("Setting databasetype to ORACLE");
 				return new OracleDbmsSupport();
 			}
-			if (PRODUCT_NAME_MSSQLSERVER.equals(product)) {
+			if (PRODUCT_NAME_MSSQLSERVER.equals(productName)) {
 				log.debug("Setting databasetype to MSSQLSERVER");
 				return new MsSqlServerDbmsSupport();
 			}
+			if (PRODUCT_NAME_MARIADB.equals(productName)) {
+				log.debug("Setting databasetype to MariaDB");
+				return new MariaDbDbmsSupport();
+			}
 		}
-		log.debug("Setting databasetype to GENERIC, productName ["+product+"]");
+		log.debug("Setting databasetype to GENERIC, productName ["+productName+"]");
 		return new GenericDbmsSupport();
 	}
 
 	public Properties getDbmsSupportMap() {
 		return dbmsSupportMap;
 	}
-
 	public void setDbmsSupportMap(Properties dbmsSupportMap) {
 		this.dbmsSupportMap = dbmsSupportMap;
 	}
+
 }